Fix parsing (#1182)
* Fix parsing of mixin types and mixin expressions in declaration

* Better detection and recovery on invalid struct initializer

---------

+++ dlang/psi-impl/src/main/kotlin/io/github/intellij/dlanguage/parser/DLangParser.kt
@@ -209,7 +209,7 @@ internal class DLangParser(private val builder: PsiBuilder) {
                 }
             }
         }
-        exit_section_modified(builder, m, DlangTypes.ALIAS_INITIALIZER, true)
+        m.done(DlangTypes.ALIAS_INITIALIZER)
         return true
     }
 
@@ -383,12 +383,12 @@ internal class DLangParser(private val builder: PsiBuilder) {
             builder.advanceLexer()
         }
         if (parseAssignExpression() == null) {
-            argumentRecover("assign expression expected")
+            argumentRecover()
         }
         m.done(DlangTypes.NAMED_ARGUMENT)
     }
 
-    private fun argumentRecover(errorMessage: String, ) {
+    private fun argumentRecover() {
         val recovery = builder.mark()
         while (!builder.eof()) {
             if (builder.tokenType === DlangTypes.OP_COMMA || builder.tokenType === DlangTypes.OP_PAR_LEFT) {
@@ -396,7 +396,7 @@ internal class DLangParser(private val builder: PsiBuilder) {
             }
             builder.advanceLexer()
         }
-        recovery.error(errorMessage)
+        recovery.error("Assign expression expected")
     }
 
     /**
@@ -4489,12 +4489,18 @@ internal class DLangParser(private val builder: PsiBuilder) {
         }
         builder.advanceLexer()
         if (!parseArgumentList()) {
+            bookmark.rollbackTo()
             m.done(DlangTypes.MIXIN_DECLARATION)
-            return true
+            return false
+        }
+        expect(DlangTypes.OP_PAR_RIGHT)
+        if (builder.tokenType !== DlangTypes.OP_SCOLON) {
+            // if no scolon, it’s probably a mixin type part of a function declaration
+            bookmark.rollbackTo()
+            return false
         }
         bookmark.drop()
-        expect(DlangTypes.OP_PAR_RIGHT)
-        expect(DlangTypes.OP_SCOLON)
+        builder.advanceLexer()
         m.done(DlangTypes.MIXIN_DECLARATION)
         return true
     }
@@ -6432,6 +6438,7 @@ internal class DLangParser(private val builder: PsiBuilder) {
         val m = builder.mark()
         expect(DlangTypes.OP_BRACES_LEFT)
         var r = true
+        var first = true
         while (!builder.eof() && builder.tokenType !== DlangTypes.OP_BRACES_RIGHT) {
             val bookmark = builder.mark()
             if (!parseStructMemberInitializer()) {
@@ -6441,12 +6448,62 @@ internal class DLangParser(private val builder: PsiBuilder) {
                 break
             }
             bookmark.drop()
-            if (currentIs(DlangTypes.OP_COMMA))
+            if (currentIs(DlangTypes.OP_COMMA)) {
+                first = false
                 advance()
+            }
+            else if (currentIs(DlangTypes.OP_COLON)) {
+                if (first) {
+                    // the first one is associative array syntax, report error for the whole block
+                    m.rollbackTo()
+                    val errorMarker = builder.mark()
+                    expect(DlangTypes.OP_BRACES_LEFT)
+                    var count = 1
+                    while (!builder.eof()) {
+                        if (builder.tokenType == DlangTypes.OP_BRACES_LEFT) {
+                            count++
+                        }
+                        else if (builder.tokenType == DlangTypes.OP_BRACES_RIGHT) {
+                            if(--count == 0) {
+                                builder.advanceLexer()
+                                break
+                            }
+                        }
+                        builder.advanceLexer()
+                    }
+                    errorMarker.error("Incorrect syntax for associative array, expected `[]`, found `{}`")
+                    return true
+                } else {
+                    // this case appeared at the middle of the struct initializer, beginning is fine
+                    // so mark only the extra as error
+                    val errorMarker = builder.mark()
+                    var count = 0
+                    while (!builder.eof()) {
+                        if (builder.tokenType == DlangTypes.OP_BRACES_LEFT) {
+                            count++
+                        }
+                        else if (builder.tokenType == DlangTypes.OP_BRACES_RIGHT) {
+                            if(--count < 0) {
+                                builder.advanceLexer()
+                                break
+                            }
+                        } else if (builder.tokenType == DlangTypes.OP_COMMA) {
+                            break
+                        }
+                        builder.advanceLexer()
+                    }
+                    errorMarker.error("Invalid syntax, `:` is not allowed here, `,` or `}` is expected")
+                    continue
+                }
+            }
             else
                 break
         }
-        expect(DlangTypes.OP_BRACES_RIGHT)
+        if (builder.tokenType !== DlangTypes.OP_BRACES_RIGHT) {
+            r = false
+        } else {
+            builder.advanceLexer()
+        }
         m.done(DlangTypes.STRUCT_INITIALIZER)
         return r
     }
@@ -7341,7 +7398,7 @@ internal class DLangParser(private val builder: PsiBuilder) {
                 }
             }
         } else if (i === DlangTypes.KW_MIXIN) {
-            if (parseMixinType()) {
+            if (!parseMixinType()) {
                 m.done(DlangTypes.BASIC_TYPE)
                 return false
             }
